Cook's Distance in Dynr

Cook's Distance in Dynr

dynr.cookdist(dataframe, id, time, ini.beta, dir = getwd())

Arguments

dataframe

a data frame object of data that contain a column of subject ID numbers (i.e., an ID variable), a column indicating subject-specific measurement occasions (i.e., a TIME variable), at least one column of observed values, and any number of covariates. The TIME variable should contain subject-specific sequences of (subsets of) consecutively equally spaced numbers (e.g, 1, 2, 3, ...). That is, the program assumes that the input data.frame is equally spaced with potential missingness. If the measurement occasions for a subject are a subset of an arithmetic sequence but are not consecutive, NAs will be inserted automatically to create an equally spaced data set before estimation. Missing values in the observed variables shoud be indicated by NA.

id

a character string of the name of the ID variable in the data.

time

a character string of the name of the TIME variable in the data.

ini.beta

the matrix of starting/fixed values for the transition matrix in the specified linear dynamic model.

dir

path for output files.

Author

Jonathan Park

Maybe we can put a description here on how Cook's distance is calculated. Also Jonathan can you give me the data set you used in the example?

I made the following changes to the code.

I changed the name to dynr.cookdist to match the other functions in the package.

I changed the argument data to dataframe to match dynr::dynr.data().

I changed the argument ID to id to match dynr::dynr.data().

I set a default value for the argument dir.

I used styler::style_pkg() to style the code for added readability.