Automatically Fit VAR(1) Models Based on Input Data

Automatically Fit VAR(1) Models Based on Input Data

dynr.var(
  dataframe,
  nv,
  id = NULL,
  time,
  dir = getwd(),
  alpha = 0.05,
  ini.mu = NULL,
  ini.cov = NULL,
  ini.beta = NULL,
  ini.loadings = NULL,
  use.mi = FALSE,
  aux = NULL
)

Arguments

dataframe

a data frame object of data that contain a column of subject ID numbers (i.e., an ID variable), a column indicating subject-specific measurement occasions (i.e., a TIME variable), at least one column of observed values, and any number of covariates. The TIME variable should contain subject-specific sequences of (subsets of) consecutively equally spaced numbers (e.g, 1, 2, 3, ...). That is, the program assumes that the input data.frame is equally spaced with potential missingness. If the measurement occasions for a subject are a subset of an arithmetic sequence but are not consecutive, NAs will be inserted automatically to create an equally spaced data set before estimation. Missing values in the observed variables shoud be indicated by NA.

nv

number of variables.

id

a character string of the name of the ID variable in the data.

time

a character string of the name of the TIME variable in the data.

dir

path for output files.

alpha

significance level for testing of transition matrix coefficients.

ini.mu

a vector of the starting or fixed values of the initial state vector.

ini.cov

a positive definite matrix of the starting or fixed values of the initial error covariance structure.

ini.beta

the matrix of starting/fixed values for the transition matrix in the specified linear dynamic model.

ini.loadings

matrix of starting or fixed values for factor loadings.

use.mi

if use.mi = TRUE, use dynr::dynr.mi() to address missing data in covariates.

aux

names of the auxiliary variables used in the imputation model if use.mi = TRUE.

Value

Returns a list of fitted VAR(1) models for each subject.

Details

I encapsulated dynrautoVAR into a temporary package so I can render the documentation and do some tests.

I made the following changes to the code.

I changed the argument data to dataframe to match dynr::dynr.data().

I changed the argument ID to id to match dynr::dynr.data().

I changed the argument p.val to alpha.

I removed the default NULL value on arguments that require explicit values (dataframe, nv, and time).

I made all initial condition arguments begin with ini..

I set a default value for the argument dir to remove it from the argument handling section.

I cleaned up the argument handling section.

I changed result$estimation.result to results$estimation.result.

I gave the output a class of dynrVar for subsequent use of the output.

I used styler::style_pkg() to style the code for added readability.

Note that I mainly based the way I documented the arguments from functions in the dynr package used within this function. If you are amenable to the way I documented this function, I will proceed to documenting the rest.

References

Add references to dynr.var() here.

Author

Jonathan Park

This function takes in a data.frame, fits individual models for each subject indicated by the ID variable and stores the fitted models as a list. Note that if ini.* arguments are not provided, initial values are set to null matrices or null vectors.